Restructured the folder paths of the repo and also introduced if statements, making the code more dynamic to fit the logic.
