--- go.mod
@@ -1,3 +1,0 @@
-module first-app
-
-go 1.21.6
